bpmn-timeline refactor
