+++ src/management-system-v2/components/bpmn-timeline/core/synchronization.ts
@@ -2,8 +2,8 @@
  * Synchronization logic for BPMN path traversal
  */
 
-import type { BPMNFlowElement, BPMNSequenceFlow } from './types';
-import { isGatewayElement } from './element-transformers';
+import type { BPMNFlowElement, BPMNSequenceFlow } from '../types/types';
+import { isGatewayElement } from '../transformers/element-transformers';
 
 export interface SynchronizationRequirement {
   requiredSources: Set<string>;
