+++ src/management-system-v2/components/bpmn-timeline/styles/BPMNTimeline.module.scss
