+++ src/management-system-v2/components/bpmn-timeline/types/types.ts
